+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Action for install tye global tool
+- Action for scaffold `tye.yaml` file into solution folder
+- Run configuration for `tye run` command